some minor fixes

+++ source/newships/shpsefn2.cpp
@@ -1127,7 +1127,7 @@ void Hook2::animate_ropeseg( Frame *space, Vector2 pos1, Vector2 pos2, int ropec
 	
 	// simulate visibility of a glittering line in the sun ...
 	double a, colscale;
-	a = atan2(dy,dx);
+	a = atan2((float)dy,(float)dx);
 	colscale = fabs(sin(a));	// flat lying = less visible.
 	
 	int col, r, g, b;
